refactor deepMerge
