+++ src/useListState/utils/index.js
@@ -1,2 +1,3 @@
 export { default as arrayToHashMap } from "./arrayToHashMap";
 export { default as throwErrorIfKeyIsNil } from "./throwErrorIfKeyIsNil";
+export { default as deepMerge } from "./deepMerge";
